Fix CPU out of range VM tests
The max_cpus was set to 12 for all of the templates I could find so the current test config wouldn't cause the error. I changed it to 16, which fixed the test.

Furthermore, there is an OutOfRangeAfterUpdate test which also failed. This was because the destroy produced the same error as the previous step and prevented it from happening. This then went on to fail the test. To fix this, I added a third step identical to the first step, to reset the config to the successful version, to then allow the destroy to proceed.