Stop panics for network interfaces that aren't connected to a network
Previously the flattenNetworkInterface function blindly dereferenced pointers to fields such as network_id without checking for nil.
However, sometimes these fields aren't set, in particular if there is a disconnected network interface attached to the VM.
In this case it is necessary to check for nil before dereferencing the pointer.

Having fixed this particular source of panic, another one occurred.
In particular, the schema.Set hash functions expected the fields to be set.
This could have been fixed by assigning an empty string in the event that the pointer was nil, instead of just not assigning anything.
However I don't believe that the hash functions are necessary at the moment as the plugin SDK does a fairly good job at hashing the set itself, and there didn't seem to be any custom behaviour required.
Also a lot of the other flatten* functions construct and return schema.Sets using these hash functions.
This also seems to be unnecessary as d.Set accepts []interface{} in place of a schema.Set so we can delegate the set hashing to that.
Therefore this commit also removes all of the use of schema.Set outside of the plugin SDK.