Fix TestAccSkytapVM_Update test by pausing between test steps
I believe the error was being caused by trying to stop the VM too soon after it had been started.
The second step attempts to rename the VM, which requires stopping it first.
This causing a lot of timeouts previously, and upon inspection in the Skytap console, it could be seen that the VM was going into a "Guest OS is not responding" error state and never coming out of it, causing the test to timeout.
Interestingly, the test didn't timeout so much when I had a lot of breakpoints and was stepping through the update more manually.
This led to the conclusion that in the normal unencumbered breakpoint-less execution, things were happening too quickly.
By increasing the "pause" delay between the steps to 1 minute, and running the tests with "-count 5" to get 5 consecutive executions, we got to a 100% pass rate so it seems to resolve the issue.
It remains yet to be seen whether this same fix applies to some of the other test failures we're seeing.