Check VM CPUs does not exceed RAM
If trying to update a VM to use more vCPUs than it has RAM, the API responds with a "409 - Conflict" error which leads the SDK to enter and endless retry loop until the context deadline expires.
To solve this, I have added a check to make sure that we are not attempting to increase the vCPUs higher than allowed.