fix: silly if err != nil bug 🤦🏻‍♂️
I attempted to fix an unhandled error from a d.Set call but neglected to include the if err != nil part.
Instead I simply always returned err, regardless of nilness, causing a premature return from the VM create function.
This, as you can imagine, was wreaking havoc on the tests.